volumedriver: introduce VolumeFwd.h
Reduce the number of objects that have to be recompiled if Volume.h
changes.

Interlude to https://github.com/openvstorage/volumedriver/issues/268
